Move the site map into Server
